models/hnn/holographic_neural_network.py
import numpy as np
import torch
import torch.nn as nn
from typing import List, Dict, Any
import json
import faiss

class EnhancedHolographicNeuralNetwork:
    def __init__(self, dimension: int = 100):
        self.dimension = dimension
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        
        # Inicializar componentes neurales
        self.encoder = nn.Sequential(
            nn.Linear(dimension, dimension * 2),
            nn.ReLU(),
            nn.Linear(dimension * 2, dimension)
        ).to(self.device)
        
        self.decoder = nn.Sequential(
            nn.Linear(dimension, dimension * 2),
            nn.ReLU(),
            nn.Linear(dimension * 2, dimension)
        ).to(self.device)
        
        # Inicializar índice FAISS para búsqueda eficiente
        self.index = faiss.IndexFlatL2(dimension)
        self.knowledge_base = []
        
        # Optimizadores
        self.optimizer = torch.optim.Adam(
            list(self.encoder.parameters()) + 
            list(self.decoder.parameters())
        )
        
        self.context_nodes = []

    def encode(self, text: str) -> np.ndarray:
        """Codifica texto en un vector holográfico."""
        # Implementar codificación de texto a vector
        vector = np.random.randn(self.dimension)  # Placeholder
        return self._normalize(vector)

    def decode(self, vector: np.ndarray) -> str:
        """Decodifica un vector holográfico a texto."""
        # Implementar decodificación de vector a texto
        return "Decoded text"  # Placeholder

    def learn(self, input_text: str, response: str) -> None:
        """Aprende de una pareja input-respuesta."""
        input_vector = self.encode(input_text)
        response_vector = self.encode(response)
        
        # Añadir al índice FAISS
        self.index.add(np.array([input_vector], dtype=np.float32))
        self.knowledge_base.append({
            'input': input_text,
            'response': response,
            'input_vector': input_vector.tolist(),
            'response_vector': response_vector.tolist()
        })

    def generate_response(self, input_text: str) -> str:
        """Genera una respuesta basada en el input y el conocimiento actual."""
        input_vector = self.encode(input_text)
        
        # Buscar vectores similares
        k = 5  # número de vecinos más cercanos
        D, I = self.index.search(np.array([input_vector], dtype=np.float32), k)
        
        # Combinar respuestas relevantes
        responses = []
        for idx in I[0]:
            if idx < len(self.knowledge_base):
                responses.append(self.knowledge_base[idx]['response'])
        
        # Por ahora, retornamos la respuesta más similar
        return responses[0] if responses else "No tengo una respuesta adecuada."

    def update_context_nodes(self, context: List[Dict[str, Any]]) -> None:
        """Actualiza los nodos de contexto."""
        self.context_nodes = context

    def export_knowledge(self) -> str:
        """Exporta la base de conocimiento."""
        return json.dumps(self.knowledge_base)

    def import_knowledge(self, knowledge_str: str) -> bool:
        """Importa una base de conocimiento."""
        try:
            knowledge = json.loads(knowledge_str)
            self.knowledge_base = knowledge
            
            # Reconstruir índice FAISS
            self.index = faiss.IndexFlatL2(self.dimension)
            vectors = [np.array(k['input_vector'], dtype=np.float32) for k in knowledge]
            if vectors:
                self.index.add(np.vstack(vectors))
            return True
        except Exception as e:
            print(f"Error importing knowledge: {e}")
            return False

    def _normalize(self, vector: np.ndarray) -> np.ndarray:
        """Normaliza un vector."""
        return vector / np.linalg.norm(vector)

    def generate_words(self, prefix: str, count: int = 10) -> List[str]:
        """Genera predicciones de palabras basadas en un prefijo."""
        # Implementar predicción de palabras
        return [f"{prefix}{i}" for i in range(count)]  # Placeholder
