requirements.txt
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
p2p-python>=0.2.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
nltk>=3.8.1
redis>=4.5.0
python-jose>=3.3.0
websockets>=11.0.3
nemo-toolkit>=1.20.0
llama-index>=0.8.0
